refactor: modernize Java code with switch expressions and pattern matching
- Replace legacy switch statements with switch expressions throughout
- Use pattern matching for instanceof checks in GameTree, BestResponse
- Replace null/collection initializations with List.of() and diamond operators
- Use Objects.equals() for string comparisons to avoid reference equality bugs
- Minor cleanups: final fields, StringBuilder, printf over println+format, 1F over Float.valueOf(1)

+++ src/main/java/icybee/solver/trainable/CfrPlusTrainable.java
@@ -68,7 +68,7 @@ public class CfrPlusTrainable extends Trainable{
     public float[] getAverageStrategy() {
         float[] retval = new float[this.action_number * this.card_number];
         if(this.cum_r_plus_sum == null || this.isAllZeros(this.cum_r_plus_sum)){
-            Arrays.fill(retval,Float.valueOf(1) / (this.action_number));
+            Arrays.fill(retval, 1F / (this.action_number));
         }else {
             for (int action_id = 0; action_id < action_number; action_id++) {
                 for (int private_id = 0; private_id < this.card_number; private_id++) {
@@ -76,7 +76,7 @@ public class CfrPlusTrainable extends Trainable{
                     if(this.cum_r_plus_sum[private_id] != 0) {
                         retval[index] = this.cum_r_plus[index] / this.cum_r_plus_sum[private_id];
                     }else{
-                        retval[index] = Float.valueOf(1) / (this.action_number);
+                        retval[index] = 1F / (this.action_number);
                     }
                 }
             }
@@ -89,7 +89,7 @@ public class CfrPlusTrainable extends Trainable{
     public float[] getcurrentStrategy() {
         float[] retval = new float[this.action_number * this.card_number];
         if(this.r_plus_sum == null ){
-            Arrays.fill(retval,Float.valueOf(1) / (this.action_number));
+            Arrays.fill(retval, 1F / (this.action_number));
         }else {
             for (int action_id = 0; action_id < action_number; action_id++) {
                 for (int private_id = 0; private_id < this.card_number; private_id++) {
@@ -97,7 +97,7 @@ public class CfrPlusTrainable extends Trainable{
                     if(this.r_plus_sum[private_id] != 0) {
                         retval[index] = this.r_plus[index] / this.r_plus_sum[private_id];
                     }else{
-                        retval[index] = Float.valueOf(1) / (this.action_number);
+                        retval[index] = 1F / (this.action_number);
                     }
                     if(this.r_plus[index] != this.r_plus[index]) throw new RuntimeException();
                     /*
