+++ src/main/java/icybee/solver/trainable/DiscountedCfrTrainable.java
@@ -94,7 +94,7 @@ public class DiscountedCfrTrainable extends Trainable{
     public float[] getAverageStrategy() {
         float[] retval = new float[this.action_number * this.card_number];
         if(this.cum_r_plus_sum == null || this.isAllZeros(this.cum_r_plus_sum)){
-            Arrays.fill(retval,Float.valueOf(1) / (this.action_number));
+            Arrays.fill(retval, 1F / (this.action_number));
         }else {
             for (int action_id = 0; action_id < action_number; action_id++) {
                 for (int private_id = 0; private_id < this.card_number; private_id++) {
@@ -102,7 +102,7 @@ public class DiscountedCfrTrainable extends Trainable{
                     if(this.cum_r_plus_sum[private_id] != 0) {
                         retval[index] = this.cum_r_plus[index] / this.cum_r_plus_sum[private_id];
                     }else{
-                        retval[index] = Float.valueOf(1) / (this.action_number);
+                        retval[index] = 1F / (this.action_number);
                     }
                 }
             }
@@ -115,7 +115,7 @@ public class DiscountedCfrTrainable extends Trainable{
     public float[] getcurrentStrategy() {
         float[] retval = new float[this.action_number * this.card_number];
         if(this.r_plus_sum == null ){
-            Arrays.fill(retval,Float.valueOf(1) / (this.action_number));
+            Arrays.fill(retval, 1F / (this.action_number));
         }else {
             for (int action_id = 0; action_id < action_number; action_id++) {
                 for (int private_id = 0; private_id < this.card_number; private_id++) {
@@ -123,7 +123,7 @@ public class DiscountedCfrTrainable extends Trainable{
                     if(this.r_plus_sum[private_id] != 0) {
                         retval[index] = Math.max(this.r_plus[index],0) / this.r_plus_sum[private_id];
                     }else{
-                        retval[index] = Float.valueOf(1) / (this.action_number);
+                        retval[index] = 1F / (this.action_number);
                     }
                     if(this.r_plus[index] != this.r_plus[index]) throw new RuntimeException();
                 }
